My Projects
Hangman: is a classic word-guessing game where a player tries to guess a word, letter by letter, within a certain number of attempts. The game begins with a random word being selected from a predefined list, and the player is presented with a series of dashes representing each letter of the word. The player can input one letter at a time, attempting to guess the word. If the guessed letter is present in the word, it is revealed in its respective position(s); if not, a part of the hangman is drawn as a penalty. The game continues until the player correctly guesses the word or runs out of attempts, resulting in the hangman being completely drawn. The player is allowed to choose whether to replay the game once it's over. The Hangman game is implemented in Python using various features such as file handling, dictionaries, sets, lists, tuples, if-elif-else statements, nested if statements, while loops, global variables, the time module, the random module, and string formatting techniques. This text-based game employs a recursive function to manage the gameplay flow and relies on user input to interact with the game.
Website grocery page: For our Information and Communication Technology (ICT) project, our group collaboratively developed an online shopping website. Within this project, I was responsible for the creation of the grocery page. Employing a blend of HTML, CSS, and a limited implementation of JavaScript, I constructed the structure, style, and interactive components of the grocery page. This included the layout and formatting of various grocery products, the visual design elements, and the integration of basic functionalities using JavaScript. My contributions were focused on ensuring an intuitive and visually appealing user interface, enabling seamless navigation and an engaging shopping experience for our website visitors.
OOP Project: This project uses C++ to create different classes for a simple graph system, implementing features like adding nodes and edges, retrieving neighbor counts, and printing graph data. It also demonstrates the use of inheritance, setter and getter functions, and runtime polymorphism. The project additionally showcases the utilization of pure virtual functions and their implementation in the derived classes. Together, these elements contribute to the development of an online shopping website.